update login pages, init register page
