+++ public/index.html
@@ -20,7 +20,7 @@
   <title>React App</title>
 </head>
 
-<body style="background-color: #121212;">
+<body>
   <noscript>You need to enable JavaScript to run this app.</noscript>
   <div id="root"></div>
