+++ src/App.js
@@ -5,15 +5,16 @@ import "./App.css";
 import { HashRouter, Navigate } from "react-router-dom";
 import { Routes, Route } from "react-router";
 
+// temp components for dev reference
+import FontsTest from "./components/fonts-test";
+
 // component migration
 import GlobalNav from "./components/navigation/nav-bar";
 import ProjectForm from "./pages/create-project";
 import APICards from "./pages/api-finder";
 import Home from "./pages/home";
-import Login from "./pages/login";
-
-// temp components for dev reference
-import FontsTest from "./components/fonts-test";
+import Login from "./pages/login-pages/login.js";
+import Register from "./pages/login-pages/register.js";
 
 function App() {
     return (
@@ -25,6 +26,7 @@ function App() {
                     <Route path="/" element={<Navigate to="/home" />} />
                     <Route path="/home" element={<Home />} />
                     <Route path="/login" element={<Login />} />
+                    <Route path="/register" element={<Register />} />
                 </Routes>
                 <ProjectForm />
                 {/* <APICards/> */}
