+++ src/pages/login-pages/login-pages.css
@@ -1,6 +1,9 @@
 .login-box {
+    background-color: #171717;
     border-radius: 10px;
     border-color: #8a8a8a;
     border-style: solid;
     border-width: 2px;
+
+    text-align: left;
 }
